Move new compilation stages to preprocessing module
